+++ spec/opal/i18next_spec.rb
@@ -172,6 +172,36 @@ RSpec.describe I18next do
     end
   end
 
+  it "can add a resource bundle" do
+    i18next.init({
+      debug: true
+    }).then do
+      i18next.add_resource_bundle("fr", "default", { french: "Français", spanish: "Espagnol" })
+      expect(i18next.get_resource_bundle("fr", "default")).to eq({ french: "Français", spanish: "Espagnol" })
+    end
+  end
+
+  it "can extend a resource bundle" do
+    i18next.init({
+      debug: true
+    }).then do
+      i18next.add_resource_bundle("fr", "default", { french: "Français", spanish: "Espagnol" })
+      expect(i18next.get_resource_bundle("fr", "default")).to eq({ french: "Français", spanish: "Espagnol" })
+      i18next.add_resource_bundle("fr", "default", { german: "Allemand" }, true)
+      expect(i18next.get_resource_bundle("fr", "default")).to eq({ french: "Français", spanish: "Espagnol", german: "Allemand" })
+    end
+  end
+
+  it "can checrk if a resource bundle exists" do
+    i18next.init({
+      debug: true
+    }).then do
+      expect(i18next.has_resource_bundle("fr", "default")).to be false
+      i18next.add_resource_bundle("fr", "default", { french: "Français", spanish: "Espagnol" })
+      expect(i18next.has_resource_bundle("fr", "default")).to be true
+    end
+  end
+
   it "can set the default namespace" do
     i18next.init({
       debug: true,